+++ src/physics/CollisionShapes.cpp
@@ -73,7 +73,7 @@ void EERIE_COLLISION_Cylinder_Create(Entity * io)
 	float d = 0.f;
 	float height = 0.f;
 	for(size_t i = 0; i < obj->vertexlist.size(); i++) {
-		if(i != (size_t)obj->origin && glm::abs(io->physics.cyl.origin.y - obj->vertexlist[i].v.y) < 20.f) {
+		if(i != obj->origin && glm::abs(io->physics.cyl.origin.y - obj->vertexlist[i].v.y) < 20.f) {
 			d = std::max(d, glm::distance(io->physics.cyl.origin, obj->vertexlist[i].v));
 		}
 		height = std::max(height, io->physics.cyl.origin.y - obj->vertexlist[i].v.y);
